requirements-dev.txt
flake8==4.0.1
mccabe==0.6.1
pycodestyle==2.8.0
pyflakes==2.4.0
#Deps do pytest
py==1.8.2
pluggy<2.0,>==0.12
attrs>=19.2.0
-r requirements.txt
#para pytest
coverage==6.3.1
pytest-cov==3.0.0
tomli==2.0.0
pytest-mock==3.7.0
pyparsing==3.0.7
packaging==21.3
